chore: add Noble support, drop Jammy support
Currently, I'm focusing more on moderizing the build towards newer
versions and it seems reasonable to focus on no more than 3 versions
including the Ubuntu LTS. Additionally, Jammy appears to lack features
even on the older 30.0.2 release.

This is tenative and probably easy to resolve in the future if support
were to be reintroduced.